backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Speech-to-Text
openai-whisper==20231117

# Alternative STT options (uncomment if needed)
# google-cloud-speech==2.21.0
# azure-cognitiveservices-speech==1.34.0

# Sentiment Analysis & Emotion Detection
transformers==4.35.2
textblob==0.17.1
vaderSentiment==3.3.2

# PyTorch Ecosystem (CPU only, matching versions)
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Async support
aiofiles==23.2.1

# Additional ML libraries (optional)
scikit-learn==1.3.2
scipy==1.11.4

# Language Processing
regex==2023.10.3

# Development tools (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mangum>=0.17.0